requirements.txt
# Healthcare Data Engineering Project Dependencies
# Python 3.9+ required
# Updated with successfully installed packages

# Database connectivity
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
greenlet==3.2.3

# Data processing and analysis
pandas==2.1.1
numpy==1.26.4
openpyxl==3.1.2
et_xmlfile==2.0.0

# Data validation
pydantic==2.4.2
pydantic_core==2.10.1
jsonschema==4.19.1
jsonschema-specifications==2025.4.1
attrs==25.3.0
referencing==0.36.2
rpds-py==0.25.1

# Environment and configuration
python-dotenv==1.0.0
PyYAML==6.0.1
annotated-types==0.7.0

# Logging and monitoring
structlog==23.1.0

# Testing framework
pytest==7.4.2
pytest-cov==4.1.0
coverage==7.9.1
iniconfig==2.1.0
pluggy==1.6.0

# Optional web interface
flask==2.3.3
streamlit==1.27.2
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0
itsdangerous==2.2.0

# Streamlit dependencies
altair==5.5.0
narwhals==1.44.0
pydeck==0.9.1
tornado==6.5.1
watchdog==6.0.0
protobuf==4.25.8
pyarrow==20.0.0
pillow==10.4.0
requests==2.32.4
rich==13.9.4
tenacity==8.5.0
toml==0.10.2
tzlocal==5.3.1
validators==0.35.0
gitpython==3.1.44
gitdb==4.0.12
smmap==5.0.2
importlib-metadata==6.11.0
zipp==3.23.0
cachetools==5.5.2

# Development and code quality
black==23.9.1
flake8==6.1.0
mypy==1.5.1
mypy_extensions==1.1.0
click==8.2.1
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.11.1
pyflakes==3.1.0
mccabe==0.7.0
Pygments==2.19.2
markdown-it-py==3.0.0
mdurl==0.1.2

# Date and time handling
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0

# JSON processing
jsonlines==4.0.0

# Security and encryption
cryptography==41.0.4
cffi==1.17.1
pycparser==2.22

# Performance monitoring
memory-profiler==0.61.0
psutil==7.0.0

# HTTP and networking
urllib3==2.5.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10

# Type hints
typing_extensions==4.14.0

# Packaging and build tools
setuptools==80.9.0
wheel==0.45.1
packaging==23.2

# Additional requirement
tabulate==0.9.0
